git-svn-id: https://svn.sourceforge.jp/svnroot/cadencii/Cadencii/trunk@596 9d680d4f-cbfa-477e-9905-73dbafb9b024

+++ bocoree/PortUtil.cs
@@ -306,11 +306,11 @@ namespace bocoree {
             byte[] dat = new byte[4];
             data = 0xffffffff & data;
             dat[3] = (byte)(data & (byte)0xff);
-            data = (byte)(data >>> 8);
+            data = (data >>> 8);
             dat[2] = (byte)(data & (byte)0xff);
-            data = (byte)(data >>> 8);
+            data = (data >>> 8);
             dat[1] = (byte)(data & (byte)0xff);
-            data = (byte)(data >>> 8);
+            data = (data >>> 8);
             dat[0] = (byte)(data & (byte)0xff);
             return dat;
 #else
